Update analyses and cluster script
Analyses.Rmd now includes a function `predict_total_simulations()` that is able to predict how many pairs of individuals I am simulating and how many total rows of data I am getting. Also updating the bash script running in the cluster as well as the R script to include a fix for the -Inf cases.